update work experience code
