+++ src/App.tsx
@@ -3,7 +3,7 @@ import '@radix-ui/themes/styles.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { ThemeProvider } from './components/ThemeProvider';
 import NavBar from './components/NavBar';
-import Work from './components/Work';
+import Work from './components/WorkExperience/WorkPage';
 import About from './components/About';
 import HomeHero from './components/HomeHero';
 import ProjectsPage from './components/Projects/ProjectsPage';
