+++ src/components/WorkExperience/workExperiences.ts
@@ -1,0 +1,51 @@
+import { WorkExperienceType } from '../../types/types';
+export const workExperiences: WorkExperienceType[] = [
+  {
+    company: 'SITE Technologies',
+    location: 'Remote (Chicago, IL)',
+    role: 'Software Engineer',
+    period: 'May 2023 - Present',
+    responsibilities: [
+      'As a Software Engineer at SITE Technologies, I contribute to frontend development efforts in a complete application rebuild using React, TypeScript, and modern frameworks. My focus is on creating intuitive user interfaces and implementing new features for our stakeholder-facing applications.',
+      'I drive code quality improvements through comprehensive documentation, code reviews, and clean code implementation while managing our codebase using git version control.',
+    ],
+    image: '/portfolio/site-tech.jpg',
+    companyUrl: 'https://www.sitetechnologies.com/',
+  },
+  {
+    company: 'SEO Scholars',
+    location: 'New York, NY',
+    role: '11th Grade Program Manager & AmeriCorps VISTA Project Manager',
+    period: 'June 2017 - April 2022',
+    responsibilities: [
+      'During my time at SEO Scholars, I managed comprehensive educational programs for 11th-grade students while serving in dual roles as Program Manager and AmeriCorps VISTA Project Manager. I developed and implemented curriculum strategies specifically tailored to meet student needs.',
+      'I fostered strong partnerships with teachers, parents, and community partners to create an effective collaborative learning environment that supported student success.',
+    ],
+    image: '/portfolio/seo-scholars.jpg',
+    companyUrl: 'https://www.seoscholars.org/',
+  },
+  {
+    company: 'Uncommon Schools',
+    location: 'Brooklyn, NY',
+    role: 'Special Projects Coordinator',
+    period: 'June 2016 - June 2017',
+    responsibilities: [
+      'As a Special Projects Coordinator at Uncommon Schools, I coordinated educational initiatives and school leadership programs while managing administrative operations. I created efficient systems to enhance overall school effectiveness.',
+      'In addition, I maintained strong communication channels between staff and families, ensuring all stakeholders remained well-informed and actively engaged in the educational process.',
+    ],
+    image: '/portfolio/uncommon-schools.jpg',
+    companyUrl: 'https://www.uncommonschools.org/',
+  },
+  {
+    company: 'Teach For America',
+    location: 'Atlanta, GA',
+    role: '10th Grade Science Teacher',
+    period: 'May 2014 - June 2016',
+    responsibilities: [
+      "At Teach For America, I designed and implemented an engaging 10th-grade science curriculum that captured students' imagination while meeting state standards. I created innovative teaching methods to improve student outcomes and academic success.",
+      'I led cross-curricular learning initiatives and built strong relationships with students and families to support their educational journey.',
+    ],
+    image: '/portfolio/teach-for-america.jpg',
+    companyUrl: 'https://www.teachforamerica.org/',
+  },
+];
